remove unused classical sim functions
